refactor: 의존 관계 개선
- DynamicHtmlProcessor 클래스가 StaticHtmlProcessor 클래스에 의존하지 않고 HttpProcessor에 의존하도록 변경
- `HttpProcessor` 공통으로 사용하는 메서드들을 추상 클래스로 이동
