+++ src/main/java/web/HttpProcessor.java
@@ -1,7 +1,10 @@
 package web;
 
+import static utils.ResourceHandler.*;
+
 import http.HttpRequest;
 import http.HttpResponse;
+import http.HttpStatus;
 
 public abstract class HttpProcessor {
     public static final String BASIC_HTTP_VERSION = "HTTP/1.1";
@@ -9,7 +12,50 @@ public abstract class HttpProcessor {
 
     abstract void process(HttpRequest request, HttpResponse response);
 
-    public void service (HttpRequest request, HttpResponse response) {
+    public void service(HttpRequest request, HttpResponse response) {
         process(request, response);
+    }
+
+    public byte[] getBytes(HttpRequest request) {
+        String extension = getExtension(request.getRequestURI());
+
+        if (FILE_EXTENSION_MAP.containsKey(extension)) { // /index.html, /min.css, ...
+            return read(RESOURCE_PATH + request.getRequestURI());
+        }
+        if (request.getRequestURI().equals("/")) { // localhost:8080/
+            return read(RESOURCE_PATH + "/" + INDEX_HTML);
+        }
+        return read(RESOURCE_PATH + request.getRequestURI() + "/" + INDEX_HTML); // /registration
+    }
+
+    public void responseHeader200(HttpResponse response, String contentType) {
+        response.setHttpVersion(BASIC_HTTP_VERSION)
+                .setStatusCode(HttpStatus.STATUS_OK)
+                .setContentType(contentType)
+                .setCharset(BASIC_CHAR_SET);
+    }
+
+    public void responseHeader302(HttpResponse response, String location) {
+        response.setHttpVersion(BASIC_HTTP_VERSION)
+                .setStatusCode(HttpStatus.STATUS_FOUND)
+                .setLocation(location)
+                .setCharset(BASIC_CHAR_SET);
+    }
+
+    public void responseMessage(HttpResponse response, byte[] resource) {
+        response.setContentLength(resource.length)
+                .setMessageBody(resource);
+    }
+
+    public void responseMessage(HttpResponse response, StringBuilder builder) {
+        response.setContentLength(builder.length())
+                .setMessageBody(builder.toString());
+
+        builder.setLength(0); // StringBuilder 초기화
+    }
+
+    public String getContentType(HttpRequest request) {
+        String extension = getExtension(request.getRequestURI());
+        return FILE_EXTENSION_MAP.getOrDefault(extension, "text/html");
     }
 }
